requirements.txt
# Core LangGraph and LLM
langgraph>=0.2.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0

# Backend
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# Frontend
streamlit>=1.40.0
requests>=2.31.0

# Utilities
PyGithub>=2.1.0
python-dotenv>=1.0.0
tenacity>=8.5.0

# Storage
sqlalchemy>=2.0
